refactor(mcp): restructure module layout and improve error handling
